Rewrite strike scoring logic
The initial assumption that doubling the strike score would work
was incorrect.  Therefore, the DoubleScoreFrame abstraction was not
workable.

In its place, I have introduced the actual logic needed to do
the calculations.

It has now become cumbersome to reason over, however, and so the
next step will be to refactor without introducing any new features.